Create README.md
In 2023/2024, I'm hoping to secure my first internship/job as a MERN stack Developer. I'm presently a freshman in college hoping to pursue my ambition of becoming a developer , but I don't want to waste my time in college. The only way I can drop out of college is to get a job in my preferred field, tech. So, if anyone is interested in constructing projects and working with me, please contact me at any moment.